Update to v7.x-2.x.
Redesigned the form_tweaks configuration format to make it easier to
use. We can now simply define groups of elements that apply to multiple
forms or roles. We can also exclude specific forms and roles from a
configuration.

The core code that implements the form_tweaks functionality has been
moved to an external file (class.FormTweaks.php) for clarity.